+++ backend/app/controllers/AnalisesTecnicas.java
@@ -1,10 +1,13 @@
 package controllers;
 
+import main.java.br.ufla.lemaf.beans.pessoa.Endereco;
+import main.java.br.ufla.lemaf.enums.TipoEndereco;
 import models.*;
 import models.geocalculo.Geoserver;
 import org.apache.commons.io.FileUtils;
 import security.Acao;
 import serializers.AnaliseTecnicaSerializer;
+import services.IntegracaoEntradaUnicaService;
 import utils.Mensagem;
 
 import javax.validation.ValidationException;
@@ -147,9 +150,9 @@ public class AnalisesTecnicas extends InternalController {
 
 		analiseTecnica.analise = Analise.findById(analiseTecnica.analise.id);
 
-		List<Notificacao> notificacaos = Notificacao.gerarNotificacoesTemporarias(analiseTecnica);
+		List<Notificacao> notificacoes = Notificacao.gerarNotificacoesTemporarias(analiseTecnica);
 
-		Documento pdfNotificacao = Notificacao.gerarPDF(notificacaos, analiseTecnica);
+		Documento pdfNotificacao = Notificacao.gerarPDF(notificacoes, analiseTecnica);
 
 		String nome = pdfNotificacao.tipo.nome +  "_" + analiseTecnica.id + ".pdf";
 		nome = nome.replace(' ', '_');
@@ -158,6 +161,30 @@ public class AnalisesTecnicas extends InternalController {
 		response.setHeader("Content-Type", "application/pdf");
 
 		renderBinary(pdfNotificacao.arquivo, nome);
+
+	}
+
+	public static void downloadPDFMinuta(AnaliseTecnica analiseTecnica) throws Exception {
+
+		verificarPermissao(Acao.BAIXAR_DOCUMENTO_MINUTA);
+
+		analiseTecnica = AnaliseTecnica.findById(analiseTecnica.id);
+
+		analiseTecnica.analise = Analise.findById(analiseTecnica.analise.id);
+
+        ParecerAnalistaTecnico parecer = ParecerAnalistaTecnico.getUltimoParecer(analiseTecnica.pareceresAnalistaTecnico);
+
+		parecer.documentoMinuta = ParecerAnalistaTecnico.gerarPDFMinuta(analiseTecnica, parecer);
+
+		parecer._save();
+
+		String nome = parecer.documentoMinuta.tipo.nome +  "_" + analiseTecnica.id + ".pdf";
+		nome = nome.replace(' ', '_');
+		response.setHeader("Content-Disposition", "attachment; filename=" + nome);
+		response.setHeader("Content-Transfer-Encoding", "binary");
+		response.setHeader("Content-Type", "application/pdf");
+
+		renderBinary(parecer.documentoMinuta.getFile(), nome);
 
 	}